Parse crew with non-regex mode, populate crew table
Crew lines are now parsed using String.split with updated separators
instead of regex to ensure that crews are in their original position
if there are empty slots.
Crew listing is now populated into the crew config, which was formerly
named RawLinePlusCrewName. This now has a TeamConfig array for each crew